rename/feat: 파일 이름 수정 및 라우터 설정
